Keystone: Fix belongsTo when validating V2 tokens
There is a mistake in the current Keystone V2 implementation, in
that the belongsTo switch is not implemented properly. It should be
an optional string, rather than a boolean, and its value added to
the `belongsTo` query field when validating tokens.

This patch modifies the validate method to implement it correctly,
and modifies the corresponding tests.

This change should have no impact on the Keystone authentication
used by the cluster, because the belongsTo query is not used.

Change-Id: Idafc9153959aaa648bad888fe0caef8d65cd9dad
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>