Fix BGP flaky tests
A new quagga sandbox image has been added, that should resolve
the flakiness of some tests. They are now marked as flaky so
they can be put on quarantine.

This image allow to supress the retransmit of non local routes
on specific quagga containers, so this is used in quagga0 to
avoid retransmit from quagga1 to quagga2 that could cause BGP
flapping.

Ref.: MI-593

Change-Id: Id3fdaf790b4395a7ee493643ed24b19c4da4254e
Signed-off-by: Miguel Ángel Herranz Trillo <miguel@midokura.com>