Adding flaky tests to quarantine CI job
To allow verification of flaky tests without disrupting gate jobs,
a new ci-quarantine.sh has been added to allow run them in a
dedicated job out of the gate. Needed options has been added to
existing CI scripts and some obsolete flaky attributes has been
removed from tests to avoid include then in quarantine by error.

Due to nose limitations, it is not possible to specify the tests
in the command line arguments and still apply filter expressions
based on attributes, so the test list has been removed from
command line and the gate tests has been marked with 'gate' attr.

The attribute can be added directly to the test function, but to
centralize the attribute assignment in just one file as it is done
now, a new test-attr.py file is used to set all the tags. Note that
the name of the file is important in order to be loaded before
the test Python files.

Additionally, the following tests has been removed from the gate
jobs (by not assigning them the gate attribute):

  test_vpnaas.py:test_ping_two_sites_two_subnets
  test_vpnaas.py:test_vpn_recreation
  test_vpnaas.py:test_admin_state_up_changes
  test_vpnaas.py:test_ipsec_container_failover
  test_vpnaas.py:test_non_vpn_subnet
  test_vpnaas.py:test_security_groups
  test_vpnaas.py:test_container_migration
  test_vpnaas.py:test_container_restored_on_agent_restart
  test_vpnaas.py:test_container_maintained_on_cluster_restart
  test_vpnaas.py:test_container_restored_on_agent_failure

Ref: MI-593
Ref: MI-685
Ref: MI-186

Change-Id: I9e33ad1e07e696ab8629c021dd3254ebc59e5e9b
Signed-off-by: Miguel Ángel Herranz Trillo <miguel@midokura.com>