MDTS: backport and enable latest changes in bgp tests
The bgp tests in 1.9 didn't contain some reliability improvements (e.g
checking for the paths to be learnt by quagga0 before sending new
traffic and waiting for the default route to be present in the router).
This patch backports the test improvements in master and fixes the
flavour to use the new quagga options to improve test reliability. This
makes it also easier to backport changes from master.

Ref.: MI-594
Signed-off-by: Xavi León <xavi.leon@midokura.com>

Change-Id: I39b515f1ad337c178f0151e7c56c2e6af5014ce9