BGP-IP: Add QuaggaContainer
This patch introduces a new container handler for the
BGP-on-interior-ports feature. All it does is send a message to the
RoutingManagerActor to start watching on the router port associated
with this container. RoutingHandler should handle the rest of the
BGP workflow: start the namespace, listen for updates to the config,
clean up the namespace.

Change-Id: I2db130e87c56afdf0ff5c7edf24c7fe68d0bdac3