MDTS: add docker logs of failed containers
We just included the `docker logs` commands of those containers that
were alive at the time mdts finished. This patch also adds the docker
logs command of the failed containers for easy debugging. As the
container is dead, we cannot execute commands on it to extract the
actual logs of the services.

Ref.: MI-700
Signed-off-by: Xavi León <xavi.leon@midokura.com>

Change-Id: I6d38302baf70f091118a4232954fa29c7f0df348