Remove VLAN ID from simulated packet
When egressing a VLAN-tagged bridge port during the simulation, we were removing
the VLAN ID from the FlowMatch, but not from the packet. Consequently we would
correctly install the flow to remove the VLAN ID from subsequent packets, but
the first packet would be emitted from user space without removing the VLAN
ID. This patch fixes the issue by removing the VLAN ID from the simulated packet
when removing it from the FlowMatch.

Signed-off-by: Brandon Berg <bberg@midokura.com>
Fixes: MI-687
Change-Id: I799acce7ef02038dd24e4b12051b25504f5ef5e8