Remove VLAN ID from simulated packet
WHen recirculating a packet for encapsulation, PushVLAN and PopVLAN actions were
being added to the flow actions, but applied to the encapsulated packet,
resulting in the VLAN ID not being pushed or popped for the first packet of each
flow. This patch applies these actions to the encaplusated packet when
appropriate.

Signed-off-by: Brandon Berg <bberg@midokura.com>
Fixes: MI-687
Change-Id: I799acce7ef02038dd24e4b12051b25504f5ef5e8