Host interfaces: remove interfaces on nsdb upon session change
Local cache of the host interface set can be out-of-sync with
the list in ZooKeeper (stored as epehemeral nodes).
It can be the case where nodes were created in a previous session,
the session timed out in the client, the agent stablished a new
session and the server side of the previous session times out
(removing the ephemeral nodes). If a physical interface is then
removed, we will try to remove it from ZooKeeper, throwing a
non existing node exception.

This patch monitors the session ID with ZooKeeper on every
host interface update. If it changed, we invalidate the local
cache, remove all interface nodes from ZooKeeper, and recreate
the zk nodes and the local cache with fresh information.
This patch also removes non-necessary reads on zk to see if the
patch actually existed (as we can now rely on the local cache
being up-to-date with the remote side).

Change-Id: Ib31275d83a2d19d6dbbb6d8a2150b62c102599da
Signed-off-by: Xavi León <xavi.leon@midokura.com>