Managers unsubscribe for deleted ref devices
This ensures that the virtual topology device managers that use
the topology prefetcher to subscribe to other referenced devices,
unsubscribe from those notifications when the underlying device is
deleted.

Also, the chain manager that subscribes to additional referenced
devices (IP address groups and jump chains) unsubscribes for the
correct identifiers when the devices are no longer references, and
unsubscribe for all devices on deletion.

Ref: MI-561
Ref: MI-672

Change-Id: I9d717d0deef12fcf2bad2d47a93b77f869c2cab8
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>