VPNaaS: Scheduler allows manual scheduling
With this change the container scheduler monitors the changes to
the container port and detects when the container port has been
manually re-scheduled or unscheduled.

When a container is unscheduled manually, by clearing the host
identifier in the container port, the scheduler will begin
rescheduling the container following the current scheduling
policy.

When a container is manually scheduled at a specific host, the
scheduler will accept the host if the host is eligible and
available. Otherwise, the scheduler will reject the manually
selected host, and will reschedule the container using the
current scheduling policy.

Test coverage: methods 76%, lines: 94%

Ref: MI-772

Change-Id: If11babf6208396736d8fc8baf14b09b2afd2f5e5
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>