BGP-IP: Add ability to assign ips to bgp namespace
bgp namespaces might service peers on multiple subnets, or on subnets
that are different from the namespace subnet. To deal with this, we
need to be able to notify the RoutingHandler of all the IPs
associated with a router, and assign those IPs to the namespace so
that traffic originating from the namespace has the correct source
ip.

Change-Id: I682004dc48739208ab5ea79ca49246bc58b1e223