MDTS: put back in override for neutron
By removing the previous neutron override to not install the FJ plugin,
we used the upstream neutron docker which has the latest available
python-midonetplugin and use the v5.x based configuration.

This patch puts only the necessary bits back in the neutron override to
configure it properly for a v1 architecture (install the local
python-midonetplugin and override the configuration without installing
any new neutron plugin package).

This fix is a partial revert for d41875ae07d43aa32c1c4cbfdb171b0ea613455e

Change-Id: I6785ed604e347a35a78ce36745bf51c01c9c905d
Signed-off-by: Xavi León <xavi.leon@midokura.com>