BGP-IP: Add arp entry updates for container ports
When spawning BGP on a container port, we need to prevent arp
from happening to the peer IPs. This is because the peer IPs will
only be reachable after going through the router, and by seeding
the arp table with the router port all bgp traffic will be directed
directly to the router.

Change-Id: I7249ea4e851da08fb38cf87db1cc965873f01041