Add ability to add Host via API
For upgrades, and to take advantage of the pre-binding mechanism where
the midonet agents bind the ports on their start-ups, allow hosts to be
created via REST API so that the upgrade orchestrator can pre-create
these hosts and its bindings and tunnel zone memberships before
upgrading the agents.  By doing so, we can create a state right before
the full agent upgrades where we get to see the final view of the
cluster data in zookeeper for verification.

Change-Id: I9cdca0181bdf4e3fdabbb5b5620289cd0797e061
Implements: MI-846