Add ability to create Host via API
The purpose of this change is to faciliate the v1.9.X -> v5.X upgrade
process. In the upgrade process, in order for v5.X agents to do the
local bindings upon their startup, and to configure tunnel zone/host
memberships during the upgrade, the hosts must be added in v5.X ZK by
the upgrade orchestration script with API instead of waiting for the
v5.X agents to add them.  This patch allows this to be possible.

This patch also adds HttpRequestChecks.scala used in the unit tests to
faciliate this and future backports.

Change-Id: I9cdca0181bdf4e3fdabbb5b5620289cd0797e061
Implements: MI-846
(cherry picked from commit 63e1f354f8eb59b7ec7eddf715a9fc6b2ae1b5d6)