Ensure chain manager fetches all current rules
Currently, when receiving an update, the chain manager compares the
old and new rule lists by size to determine whether it needs to
fetch any new rules. If the sizes match, the manager assumes that
it has all rules and does not attempt to fetch new ones.

This change ensures the manager compares old and new rule sets by
content, such that all new rules are fetched.

Ref: MI-598

Change-Id: I15425ed0cf4f366d773c714b22da5000a7103e88
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>