Fix flake in HaproxyTest
There was a race in shutdown because when the host was being deleted,
the haproxy monitor would have modified deleted a port which referenced
the host, thus modifying the host. In production, this would be avoided
using the topology lock, but this was only mocked in the test.

I've changed the mock to use a ReentractLock instead, so now it works
more like it would in production, and also fixes the race.

Change-Id: Iebe7105150053490c06595619f7a02337e2e564d
Signed-off-by: Ivan Kelly <ivan@midokura.com>