Increase load in perf test, decrease running time
Increases messages per second from 10 to 100. Decreases running time
from 1 hr to 30 minutes. Also allow up to a minute for traffic to
arrive, as, if the job does get into GC trouble, we don't want it to
die. No GC should take longer than a minute.

The job had been running with debug logs on previously, which distorted
the results.

Change-Id: I1fbb7d876e8767df8d2b06428115a2e6da58e1a0
Signed-off-by: Ivan Kelly <ivan@midokura.com>