FlowState: introduce cluster minions on the agent
This patch enables the agent to start a subprocess of the Midonet
Cluster component with a subset of the minions enabled. The intent of
these minions is to move some of the complexity of the current agent
into an external process, helping to alleviate memory pressure on the agent
fast path. For that purpose, we introduce a new cluster template with
a subset of the minions meant to be executed by the agent. Only the flow
state minion is enabled right now.

Ref.: MI-656
Change-Id: I53d07c35e80caec9f3dae6c3fd7e6631513512c4
Signed-off-by: Xavi León <xavi.leon@midokura.com>