FlowState: send state to flow state minion instead of cassandra
This patch removes the dependency on cassandra to store
flow state messages, removing the cassandra client from the fast path.
We still fetch keys from cassandra upon port migration and agent reboot,
much less frequent events than sending flow state for each simulation.
This patch now sends flow state messages to a the flow state minion on
the same agent. For that, we send a UDP packet to a configured port on
localhost with the flow state message so that the agent minion can 
receive it and send it to cassandra, reducing the memory pressure on
the agent fast path.

Ref.: MI-656
Change-Id: I033cd175e862f42a32238b76d356836d0e8dea5f
Signed-off-by: Xavi León <xavi.leon@midokura.com>