MDTS: fix flaky new interface test
The test_new_interface_becomes_visible is prone to flakeyness because
the previous test reboots the agents. In some cases, the agent reports
its host status as 'alive' even though it didn't finished its boot
sequence and the check for the new interface fails. This patch retries
several times for the new interface to become visible on the api before
giving up.

Change-Id: Iaa64225140abd90857cdf21a1f7a251f815a14a1
Signed-off-by: Xavi León <xavi.leon@midokura.com>