Update patch set 43
Patch Set 43:

We cannot bind to the 6677 port on the agent because it's already bound by the agent for the vxlan tunneled packets (used to send flow state). So we need to bind to another port and just send regular UDP packets to that place. Given this, it's may be just as easy to communicate with pipes as Alex suggested. We can change it in another patch.

Patch-set: 43
Label: Code-Review=0