Agent: Introduce agent minions
This patch introduces the agent minions, similar to what already exists
on the cluster. The purpose of this framework is to allows abstract
services to be instantiated and executed on demand and on a separate
process, reducing the memory presure on the agent fastpath. The current
implementation does not contain any service implementation yet, which
will be introduced in future patches. Agent minions must implement the
Minion abstract class and annotate the implementation with the
MinionService annotation so it can be discovered and automatically
started if configured to do so.

Ref.: MI-656
Signed-off-by: Xavi León <xavi.leon@midokura.com>

Change-Id: Ib4651b142cc970eecf240ab3de50c95ca626cc4a