Disallow concurrent nat block requests in same range on a single agent
If there are multiple packets being processed and each needs a snat
source port allocated, and there's no nat block available, each packet
will trigger a new block to be allocated until the agent gets one block
allocated.

This cause overallocation of blocks for the agent. In the pathological
case, where no blocks are free, so the least recently used block is
used, each packet will end up trying to request the same block, causing
the agent to race with itself for the block. This is highly wasteful of
zookeeper resources.

This fix prevents there from being more than one outstanding block
request for a nat range, thereby reducing the traffic to zookeeper,
avoiding overallocation and making the time to a block being available
more predictable.

Ref MI-898

Change-Id: I4a2e0b9d597b1b1241b432d5c12336ce0664e1da
Signed-off-by: Ivan Kelly <ivan@midokura.com>