Minions: annotate minions with node type
Currently, we discover minions by it's package namespace. That may force
external users on a specific package organization, which may not be
desirable. This patch removes the package namespace restriction and
includes a new parameter to the MinioniService annotation to specify
where a specific minion should run (AGENT or CLUSTER).

Change-Id: I6c1c27184e44174a2a0520953026cb5b89da7832
Signed-off-by: Xavi León <xavi.leon@midokura.com>