Add RejectedExecutionHandler to agent executors
Currently, when shutting down the agent, the agent will throw an
exception and stop if tasks are still pending on the VT executors
queue, since the default policy is an abort policy.

This change adds a "caller-runs" rejected execution handler to the
VT executor, with which: aborted tasks are executed immediately on
the caller thread if the executor is not shutting down, or
discarded if the executor is shutting down. This should prevent
such failures when shutting down the agent.

The methods for creating the new executors are groups in a new
utility companion object, that can later be used in other places.

Change-Id: I4dfd4295f05ae3a4f7fc7da288e391aab5e4b653
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>