Host request proxy handles port failures
This change modifies the HostRequestProxy to handle failures
retuned by the VirtualTopologyActor when loading the data for a
port binding. If a port cannot be loaded (because it cannot be
read or is deleted) it will marked as such in an internal cache
with the host request proxy, and ignore on future updates for
the same host.

The change intendes to prevent the proxy of looping continuously
on requesting the binding information for ports that are no longer
available.

Note, that any host update that removes a port binding will clear
the state for those ports in the local cache, such that they can
be re-added in the future.

The patch also adds unit tests for the host-request proxy.

Change-Id: Ic4653c96cea4aaeb54a40e269a4eea30fadb9c8b
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>