Packaging: enable deb/rpm to be compiled on same gradle run
Right now we reuse the temporal directory and fpm options
when building packages on a given module. This makes it
impossible to build several packages for different package
targets on the same gradle run (as the second run will have
fpm option from the first one). Also, the cleanPkg task is
only run once which prevents us from reusing the temporal
directory.

This patch separates the workflow to generate the packages
on different tasks so they can all be executed depending
on the targets specified when calling gradle. This is in
preparation for the platform9 packaging which is a metapackage
containing both deb and rpm packages.

Change-Id: Ic92145119a4d0f3898da121b50ac5cab8c4a45a7
Signed-off-by: Xavi León <xavi.leon@midokura.com>