Partial transaction failure leaves local VTEP cache in inconsistent state
When updating a MAC value on the VTEP table, there where no retries.
This meant that if a Physical locator needed to be added but failed
before adding the new MAC, that would generate an error when trying to
insert the new MAC, since the Physical locator associated with it failed
to get saved on the VTEP.
To fix this we add a series of retries until we get a success response
on the transaction.

Change-Id: If87acfa652c001f54c310847f115c7ac67b637aa
Ref: MI-728
Signed-off-by: Mateo Hermida <mateo@midokura.com>