ZOOM: Clear terminated observables
ZOOM maintains a cache of requested object or class observables,
such that subsequent requests for the same objects/classes do
not create a new observable instance backed by a new ZooKeeper
watcher, but rather reuse the same observable that was exposed
earlier.

This patch fixes an issue where the terminated observable are
never removed from this cache, preventing them to be collected
by the garbage collector.

Fixes: MI-860

Change-Id: Icc4f561803a45b2244a744fe14accfd3f9b2b123
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>