FlowState: send msg to minion even if peer list is empty
We always send flow state to the agent minion for it to send it to
Cassandra. However, when the peer list was empty, we didn't build the
flow state message on the packet context so we sent an empty message to
the minion, meaning that the associated flow state was lost. This
happens when the ingress and egress ports for a given packet belongs to
the same agent.

This patch ensures that we build the flow state message regardless of
the peers we need to send it to and adds a regression test.

Ref.: MI-913
Signed-off-by: Xavi León <xavi.leon@midokura.com>

Change-Id: I80e60d6d24a586a9263dd811d12ad86577eb437f