Randomize ports on API tests for all API tests
API tests suffered from flakeyness on the CI because tests are executed
in parallel and the jersey framework was initialized to use the same
random port for all tests using a system wide property.

This patch makes all API tests inherit from a common parent that
overrides the method that actually sets the port on the jersey test
framework. On every single test, a new random ephemeral port will be
used (by binding to port 0) with a retry mechanism to ensure it's not
already used on the host.

Ref.: MI-939
Signed-off-by: Xavi León <xavi.leon@midokura.com>
Signed-off-by: Miguel Ángel Herranz Trillo <miguel@midokura.com>

Change-Id: I6119c306b2102c2d9e7fd36f837bdf64b0a009a2