ZOOM: Add close handler to NodeObservable
Currently, there is no mechanism for an external observer to know
whether a NodeObservable is closed except subscribing to it.
However, subscribing to a NodeObservable prevents it to close
until all subscribers have unsubscribed.

This patch adds an out-of-band mechanism for the NodeObservable
to notify that it has closed its underlying watcher to storage,
either because it has no more subscribers or because the object
it watches has been deleted (when `completeOnDelete` flag is
true).

This will be used in a subsequent patch to remove closed
observable from the ZOOM observable cache.

Fixes: MI-860

Change-Id: Ib54514f6b4250581f0d1c291875a96f8aaf96e1f
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
(cherry picked from commit d71508356592b72481af78d8b6a6d6a68076f12d)