Update patch set 1
Patch Set 1:

No, it won't from what I see. If there's flow state being generated, we need to send it to the minion to support reboots (even if there's no peers on the list). 

However, we only reach that code path if the PacketWorkflow filled the conntrackTx and natTx state tables before. If there's no state being generated, we don't send any flow state message to the minion (nor to any of the peers).

Patch-set: 1
Label: Code-Review=0