Randomize ports on API tests using 0 port as value
Since gradle launch several processes, it turns out that generating
a JVM local thread-safe sequence of ports is not enough to avoid
race conditions among different processes.

This patch binds to 0, which makes the OS to use a random ephemeral
port. This ensures that the port we provide to the jetty test
framework is actually free.

Ref: MI-939

Change-Id: I559bca38eefe27d87aeb7f180a5177674c062055
Signed-off-by: Miguel Ángel Herranz Trillo <miguel@midokura.com>
Signed-off-by: Xavi León <xavi.leon@midokura.com>