Randomize ports on API tests using 0 port as value
Since gradle launch several processes, it turns out that generating
a JVM local thread-safe sequence of ports is not enough to avoid
race conditions among different processes.

So instead of that, the getPort method in RestApiTestBase has been
changed to return just a 0 value, that should be used in subsequent
calls to socket libraries, indicating that a random ephemeral port
should be used.

Ref: MI-939

Change-Id: I559bca38eefe27d87aeb7f180a5177674c062055
Signed-off-by: Miguel Ángel Herranz Trillo <miguel@midokura.com>