Allow multiple sandbox to run on the same machine
Midonet sandbox allows sandboxes to have different names, but mdts tests
assume the sandbox will be named "mdts". This does not allow us to run
multiple sandbox environments on a same machine. With this change the
caller of the run_tests.sh scripts will be allowed to change a value in
mdts.conf with the name of the sandbox in which the tests will be
performed. The provisioning scripts' pipework and bridge setups now also
take into account the sandbox name so as to avoid using the same
interfaces in different sandboxes.

Note: For legacy support in existing scripts, the default name for
a sandbox is set as "mdts" in the mdts.conf file.

Change-Id: I4485eadc2f61153d413d6863ca922f397486e1ea
Signed-off-by: Mateo Hermida <mateo@midokura.com>