Bridge handles ARP requests for pre-seeded IPs
This change ensures that a bridge replies to an ARP request for an
IP found in the bridge's IP-MAC table, without forwarding the
packet. This fixes an issue where ARPs addresses to a peer router
port would be forwarded to the router, even if the bridge had that
IP already preseeded in its IP-MAC table.

Fixes: MI-950

Change-Id: I37ccbbd15b241ea8ccf49c8ce23624a26e1a18af
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
(cherry picked from commit 33f7c54b2cf2389bc1b4784412bbfdad07aa0f60)