Let Jetty honor the "X-Forwarded-Proto" header
Accessing the MidoNet API through an SSL enabled proxy (that itself
connects to the API via plain HTTP) fails.

Reason is that the Jetty server sends a redirect (302) to an HTTP URL
(instead of HTTPS).

This patch lets Jetty honor the "X-Forwarded-Proto" header sent by the
proxy so that URLs will use HTTPS.

Change-Id: I2b4b258870ce26bf280b179558e39f87bb79c19d
Fixes: https://midonet.atlassian.net/browse/MNA-1149
Signed-off-by: Jan Hilberath <jan@midokura.com>