Let Jetty honor the "X-Forwarded-Proto" header
Accessing the MidoNet API through an SSL enabled proxy (that itself
connects to the API via plain HTTP) fails because the Jetty server
sends a redirect to an HTTP URL (instead of HTTPS).

This patch adds the "ForwardedRequestCustomizer" to the Jetty
configuration. This customizer looks at HTTP request for headers that
indicate it has been forwarded by one or more proxies.

Specifically handled are:

  X-Forwarded-Host
  X-Forwarded-Server
  X-Forwarded-For
  X-Forwarded-Proto

If these headers are present, then the Request object is updated so
that the proxy is not seen as the other end point of the connection
on which the request came

Change-Id: I2b4b258870ce26bf280b179558e39f87bb79c19d
Fixes: https://midonet.atlassian.net/browse/MNA-1149
Signed-off-by: Jan Hilberath <jan@midokura.com>