MidoConfTool: default to json output
Customers find it cumbersome and not consistent to parse HOCON
output, which only quotes fields if they are alphanumeric and
leaves the rest unquotted. This patch enables JSON by default
(quotes everything) and also introduces a new switch for the
dump and get commands to optionally output also the descriptions
associated to the configuration keys (disabled by default for
a plain json output).

Ref.: MI-834

Change-Id: Id7d53bff98fa89a39af715146e990efac2d1a592
Signed-off-by: Xavi León <xavi.leon@midokura.com>