Update patch set 7
Patch Set 7:

Also, in neutron_translation.md, mention the port group created on router creation to maintain the list of router interface ports so that we can create a FWaaS rule that applies only to packets ingressing these ports.  Also that router interface port is added to this port group.

Patch-set: 7