SSH connection to MDTS targets
Currently MDTS use the a python Docker API to access the
containers. This change allow MDTS to use a SSH connection instead,
this way MDTS can be used against other machines providing the
services to be tested.

To use it a YAML file called ssh-containers.yaml must be present in
the working directory of ./run_tests.sh. If present, SSH is used
instead of Docker API for all the services/containers.

The file is the equivalent version of a Sandbox flavour Dockerfile,
containing the "containers" names, API interfaces and connection
details.

A sample YAML file can be generated from a running sandbox by using

    tests/ssh-containers-from-docker.sh

script, that generates this file in the current directory.

The SSH credentials have to be provided system-wide or be provided
explicitely by overriding the following optional default entry per
each container:

    SSHSettings: {User: 'root', Password: 'docker.io'}

This is the default user/password for the ubuntu-upstart:14.04 based
midonet-sandbox images.

WARNING: not all test suites has been tested using this method, so
some of them may fail in this mode. This can be due to wrong escape
incompatibilities between the python Docker API and the python SSH
API (paramiko). Only the following tests has been run in this mode
against a sandbox:

   functional_tests  test_fip_icmp
                     test_brige:test_mac_learning
                     test_brige:test_dhcp
   performance_tests test_perf_flowstate

Ref: MI-896

Change-Id: Iaf46715a5e3ec5a02d2a5563d8b23d5fab7d6fa5
Signed-off-by: Miguel Ángel Herranz Trillo <miguel@midokura.com>