SSH connection to MDTS targets
Currently MDTS use the a python Docker API to access the
containers. This change allow MDTS to use a SSH connection instead,
this way MDTS can be used against other machines providing the
services to be tested.

See mdts.conf and example-ssh-containers.yaml for details

WARNING: not all test suites has been tested using this method, so
some of them may fail in this mode. This can be due to wrong escape
incompatibilities between the python Docker API and the python SSH
API (paramiko). Only the following tests has been run in this mode
against a sandbox:

   functional_tests  test_fip_icmp
                     test_brige:test_mac_learning
                     test_brige:test_dhcp
   performance_tests test_perf_flowstate

Ref: MI-896

Change-Id: Iaf46715a5e3ec5a02d2a5563d8b23d5fab7d6fa5
Signed-off-by: Miguel Ángel Herranz Trillo <miguel@midokura.com>