Agent: Use the correct variables for VPNaaS IPSec
For ipsec.conf: replace hard-coded values for ike= and phase2alg= with
the real settings from the IPSec policies, use the configured remote
peerId instead of the peer's IP address. This is a required field in the
OpenStack VPNaaS configuration and can be the same as the peer's IP
address. Ideally we would use the local peerId also if it was
configurable... For ipsec.secrets: use the remote peerId instead of the
peer's IP address.

Change-Id: I31e23dba112fd57e051ccccae22f5b96af30ab15
Ref: MNA-1150
Signed-off-by: syoder <syoder@imagestream.com>
(cherry picked from commit ea211303c19c41d127d0a94f3e89a02de42386f2)