Fix wrong default value for Keystone logger
Instead of a log level (DEBUG, INFO, ...), the Keystone logger
"cluster.loggers.org.midonet.cluster.auth.keystone" contains
its description as default value.

This is caused by a typo where the description is defined with
the same name as the log level and thus overwrites the log level
definition.

This patch corrects that.

Change-Id: I41d6b991df5266282bffd6538cc063735b36df2a
Fixes: https://midonet.atlassian.net/browse/MNA-1151
Signed-off-by: Jan Hilberath <jan@midokura.com>