ZOOM: Add close handler to PathDirectoryObservable
Currently, there is no mechanism for an external observer to know
whether a PathDirectoryObservable is closed except subscribing to
it. However, subscribing to such an observable prevents it to close
until all subscribers have unsubscribed.

This patch adds an out-of-band mechanism for the directory
observable to notify that it has closed its underlying watcher to
storage, either because it has no more subscribers or because the
object it watches has been deleted (when `completeOnDelete` flag
is true).

This will be used in a subsequent patch to remove closed
observable from the ZOOM observable cache.

Fixes: MI-860

Change-Id: Id64b306b0cff4ee45c1038c676b1cc216cc4ca50
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
(cherry picked from commit 61dac9af9c47cef74eef86f6db5c7c13b56ab253)