ZOOM: Clear closed state observables
ZOOM state maintains a cache of requested single and multi key
observables, such that subsequent requests for the same state
keys do not create a new observable instance backed by a new
ZooKeeper watcher, but rather reuse the same observable that
was exposed earlier.

To address concurrency, any observable upon completion notifies
an `observable closed` error such that any observer that still
holds a reference to an observable that closed its ZooKeeper
watcher can recover to a new fresh observable. ZOOM handles
this recovery internally, such that when a new observable is
needed, it is automatically added to the observables cache (or
one from the cache is used in the case of a concurrent request).

However, currently, cached observables are only cleared when
they are replaced by a fresh observable, and therefore ZOOM will
keep references indefinitely to observables that have already
terminated or have no more subscribers.

Furthermore, in the current code, ZOOM will request and add a
new observable to the cache, even if this observable is not
immediately subscribed.

This patch addresses this issues as follows:

* When requesting a state key observable, ZOOM will create and
  cache an underlying NodeObservable only when the returned
  observable is being subscribed. This will prevent keeping
  references to observables that may not ever be subscribed.
  Note that such observables will never emit any notification,
  nor call their close handler. This new approach will create
  a new Observable instance for every call of the `observable`
  method, but it has the advantage that the storage does not
  hold any references to it and it can be GCed when no longer
  needed.

* For a state key observable, it adds a close handler to the
  underlying NodeObservable, that will remove the observable
  from the cache.

Change-Id: Ie36ffa7c654fa1878b7802554ff86fc9a78cad32
Fixes: MI-860
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
(cherry picked from commit c198a06f8da85a52eb942997e3c02fe33a16bb03)