Add missing type parameters to Neutron Zoom Plugin
The Neutron Zoom Plugin was missing type parameters for the
getSecurityGroup() and getSecurityGroupRule() functions.

As a result, when calling these functions a NullPointerException
was thrown in the API and HTTP error 500 returned to the client.

Change-Id: Ifad2cf2e55cc5f6a4bda2521fdf8f7b8f429deca
Fixes: https://midonet.atlassian.net/browse/MNA-1144
Signed-off-by: Jan Hilberath <jan@midokura.com>
(cherry picked from commit 9ed23a56c1081ce1c7f11f6fb17972c8aa96f708)