BGP-IP: define port id in learned routes
When the routing handler creates learned routes it will set the port
in the key to the next hop port in the route. This can be an issue
for BGP on interior ports, because interior ports do not have a host
id associated with them. By allowing the port ID to be specified in
the key, we can always choose a port that has a host id.

Change-Id: I005c5b289b4997f4231e1e94984fe1dde68a087a