10kp: A common trait for encoding paths
This adds a trait for encoding paths used by the state tables/
replicated map classes, such that the trait can be mixed-in
without requiring the implementation of the encode/decode
methods and therefore eliminating this duplication of code.

It also moves this encoding to the NSDB module, and provides
static instaces for the encoding of MAC-UUID, MAC-IPv4 and
IPv4-MAC tables.

Ref: MI-1003

Change-Id: I2c931f2632ded3df2167978ad99db507dcc0c43e
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>