Add tunnel keys to the port state storage
This change adds a new Protocol Buffers message to represent the
state of a port, which indicates the host where the port is active
and the agent-local tunnel key for this port. The patch also
modifies the port state storage to add a utility wrapper over the
ZOOM storage to easily read and write the port state.

This changes the data written to the port ACTIVE key from the
a single string representing the host UUID to a JSON represenation
of the Protocol Buffers message.

Backward compatibility is assured in two ways:

* For reads at the agents of previous versions, the change has no
  effect, because the value of the state key was not used, only
  whether is was set (port active) or not (port inactive).

* For reads at the agents 5.2+, the port state storage wrapper
  support both old format (only host ID, no agent-local tunnel key)
  and the new format (host ID and agent-local tunnel key).

Ref: MI-112

Change-Id: I4e2aca81c8be098573ee76745e483ac502306acd
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>