Update patch set 5
Patch Set 5: Code-Review+2

It would be nice to add a test for this, but I don't know of any easy way to do so as it requires populating learned route table.  Brandon, let me know if you can think of a good way so that we can add tests for this.

Patch-set: 5
Label: Code-Review=+2