DPE allocates local tunnel keys
Datapath triads add a new local tunnel key, and renames the
previous tunnel key as a remote tunnel key (being allocated by
the cluster). The local tunnel keys are allocated for every new
datapath triad corresponding to a new network interface, and
release when the interface is shutdown and the triad no longer
used.

Both local and remote tunnel keys are mapped to the datapath
triads. In addition, the local tunnel keys are reported to the
virtual-to-physical-mapper to be updated in the local port's
state.

Change-Id: I8728bca01fe39d3c94ece8f35771b00ad9a878c9
Ref: MI-112
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>