Midonet scala interface for 6wind
This change add a new flow writer called SixwindWriter that
interface the native mido6wind library (see MI-976).

If the library cannot be loaded, it logs the error and ignore
any subsequent write.

If the library can be load, but the processing of a flow return
an error, it is logged with the error code.

Ref: MI-977
Ref: MI-976

Change-Id: I0920be6765676803e27875f5d8bd3658599cd514
Signed-off-by: Miguel Ángel Herranz Trillo <miguel@midokura.com>