MDTS: set transit quaggas as passive
Our BGP topology consists of a master bgp speaker (quagga0) peering with
two transit bgp speakers (quagga1 and quagga2) that peer each with two
agents (midolman1 and midolman2). All of them were set as active
speakers which could create a connection conflict leading to some
connections never get established.

This patch leverages a recent patch on sandbox to set the transit bgp
speakers as passive, and letting the bgp speakers on the agents to
always initiate and establish the connections whenever a new virtual bgp
router is set up on the virtual topology.

Ref.: MI-186
Signed-off-by: Xavi León <xavi.leon@midokura.com>

Change-Id: Ie01d99e842fa330da4a6c931378d8cb9d2859630